requirements.txt
# Core dependencies
streamlit
pandas
numpy
plotly
websocket-client
requests

# Analytics and statistics
statsmodels
scipy
scikit-learn

# Database
sqlalchemy

# Utilities
python-dotenv
pydantic

